+++ apps/demo/src/types/Bindings.ts
@@ -1,0 +1,7 @@
+export type Bindings = {
+  JWKS_URL: string;
+  JWKS_SERVICE: {
+    fetch: typeof fetch;
+  };
+  AUTH_URL: string;
+};