+++ packages/kysely/src/migrate/migrations/2024-11-18T10:37:00_act_as.ts
@@ -1,0 +1,13 @@
+import { Kysely } from "kysely";
+import { Database } from "@authhero/kysely-adapter";
+
+export async function up(db: Kysely<Database>): Promise<void> {
+  await db.schema
+    .alterTable("logins")
+    .addColumn("authParams_act_as", "varchar(255)")
+    .execute();
+}
+
+export async function down(db: Kysely<Database>): Promise<void> {
+  await db.schema.alterTable("logins").dropColumn("authParam_act_as").execute();
+}