+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @authhero/demo
 
+## 0.1.0
+
+### Minor Changes
+
+- Get the demo project rendering
+
+### Patch Changes
+
+- Updated dependencies
+  - authhero@0.10.0
+  - @authhero/kysely-adapter@0.18.0
+
 ## 0.0.36
 
 ### Patch Changes