+++ packages/authhero/CHANGELOG.md
@@ -1,5 +1,16 @@
 # authhero
 
+## 0.10.0
+
+### Minor Changes
+
+- Get the demo project rendering
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.22.0
+
 ## 0.9.0
 
 ### Minor Changes